Refactor: Implement Clean Architecture and Modular CSS
