+++ src/features/report/styles/dashboard.css
@@ -1,27 +1,27 @@
 /* Layout Wrappers */
-.app-wrapper { 
-  min-height: 100vh; 
-  width: 100%; 
-  padding: 30px 16px; 
-  box-sizing: border-box; 
-  display: flex; 
-  justify-content: center; 
+.app-wrapper {
+  min-height: 100vh;
+  width: 100%;
+  padding: 30px 16px;
+  box-sizing: border-box;
+  display: flex;
+  justify-content: center;
   align-items: flex-start;
   background: var(--gradient-app) !important;
   background-attachment: fixed !important;
 }
 
-.form-card { 
-  width: 100%; 
-  max-width: 580px; 
+.form-card {
+  width: 100%;
+  max-width: 580px;
   background: var(--color-bg-card);
   backdrop-filter: blur(20px);
-  padding: 32px 28px; 
-  border-radius: var(--radius-xl); 
+  padding: 32px 28px;
+  border-radius: var(--radius-xl);
   box-shadow: var(--shadow-card), 0 0 0 1px rgba(255, 255, 255, 0.8) inset;
-  box-sizing: border-box; 
+  box-sizing: border-box;
   border: 1px solid var(--color-border-subtle);
-  position: relative; 
+  position: relative;
 }
 
 /* Card Top Header */
@@ -103,32 +103,32 @@
 }
 
 /* Section Heading & Step Badges */
-.section-heading { 
+.section-heading {
   display: flex;
   flex-direction: row;
   align-items: center;
   gap: 8px;
-  color: #86198f; 
-  font-size: 13px; 
-  font-weight: 800; 
+  color: #86198f;
+  font-size: 13px;
+  font-weight: 800;
   margin-bottom: 10px;
 }
 
-.section-heading > span:first-child { 
-  background: linear-gradient(135deg, #d946ef 0%, #9333ea 100%); 
-  border: none; 
-  color: #ffffff; 
-  width: 24px; 
-  height: 24px; 
-  display: inline-flex; 
-  align-items: center; 
-  justify-content: center; 
-  border-radius: 6px; 
-  font-size: 12px; 
-  font-weight: 800; 
-  flex-shrink: 0; 
-  box-shadow: 0 2px 8px rgba(217, 70, 239, 0.35); 
-  transition: all 0.3s; 
+.section-heading > span:first-child {
+  background: linear-gradient(135deg, #d946ef 0%, #9333ea 100%);
+  border: none;
+  color: #ffffff;
+  width: 24px;
+  height: 24px;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 6px;
+  font-size: 12px;
+  font-weight: 800;
+  flex-shrink: 0;
+  box-shadow: 0 2px 8px rgba(217, 70, 239, 0.35);
+  transition: all 0.3s;
   line-height: 1;
 }
 
@@ -159,8 +159,8 @@
   font-size: 10px;
   font-weight: 800;
   margin-left: 6px;
-  display: inline-flex; 
-  align-items: center; 
+  display: inline-flex;
+  align-items: center;
   justify-content: center;
   animation: pulseBadge 2s infinite cubic-bezier(0.66, 0, 0, 1);
   height: 18px;
