+++ src/features/report/styles/form-responsive.css
@@ -1,0 +1,200 @@
+/* Mobile Responsive Adjustments - RECOVERED UTUH */
+@media (max-width: 600px) {
+  .app-wrapper {
+    padding: 12px 10px !important;
+    margin: 0 !important;
+    min-height: 100vh !important;
+    align-items: flex-start !important;
+  }
+
+  .form-card {
+    width: 100% !important;
+    max-width: 100% !important;
+    padding: 18px 14px 24px 14px !important;
+    border-radius: 18px !important;
+    border: 1px solid var(--color-border-subtle) !important;
+    background: var(--color-bg-card) !important;
+    box-shadow: 0 10px 30px rgba(0, 0, 0, 0.2) !important;
+  }
+
+  .card-top-header {
+    margin-bottom: 14px !important;
+    padding-bottom: 12px !important;
+    gap: 8px !important;
+  }
+
+  .card-top-header img {
+    width: 36px !important;
+    height: 36px !important;
+    border-radius: 8px !important;
+  }
+
+  .form-title {
+    font-size: 15px !important;
+    margin-bottom: 0 !important;
+    font-weight: 800 !important;
+  }
+
+  .card-top-header span {
+    font-size: 11px !important;
+  }
+
+  .sheets-header-btn {
+    padding: 6px 10px !important;
+    font-size: 11px !important;
+    border-radius: 10px !important;
+    gap: 4px !important;
+  }
+
+  .sheets-header-btn svg {
+    width: 13px !important;
+    height: 13px !important;
+  }
+
+  .staff-tabs-wrapper {
+    margin-bottom: 14px !important;
+    padding: 3px !important;
+    border-radius: 12px !important;
+  }
+
+  .staff-tab-button {
+    font-size: 12px !important;
+    height: 38px !important;
+    padding: 4px 8px !important;
+    border-radius: 9px !important;
+  }
+
+  .global-section,
+  .active-form-section,
+  .draft-list-wrapper {
+    padding: 14px 12px !important;
+    gap: 12px !important;
+    margin-bottom: 14px !important;
+    border-radius: 14px !important;
+  }
+
+  .section-heading {
+    display: flex !important;
+    flex-direction: row !important;
+    align-items: center !important;
+    font-size: 12px !important;
+    margin-bottom: 8px !important;
+    margin-left: 0 !important;
+    gap: 8px !important;
+  }
+
+  .section-heading > span:first-child {
+    width: 22px !important;
+    height: 22px !important;
+    font-size: 11px !important;
+    border-radius: 6px !important;
+    display: inline-flex !important;
+    align-items: center !important;
+    justify-content: center !important;
+    flex-shrink: 0 !important;
+    line-height: 1 !important;
+  }
+
+  .form-label {
+    font-size: 11px !important;
+    margin-bottom: 4px !important;
+  }
+
+  .custom-select-trigger,
+  .form-input {
+    padding: 0 12px !important;
+    font-size: 12px !important;
+    height: 42px !important;
+    border-radius: 10px !important;
+  }
+
+  .file-dropzone {
+    padding: 14px 10px !important;
+    border-radius: 12px !important;
+  }
+
+  .submit-button,
+  .btn-add-draft {
+    height: 44px !important;
+    font-size: 12px !important;
+    border-radius: 10px !important;
+  }
+
+  .settings-active-banner {
+    padding: 10px 12px !important;
+    border-radius: 12px !important;
+    gap: 8px !important;
+  }
+
+  .sender-pill-badge,
+  .date-pill-badge {
+    padding: 5px 10px !important;
+    font-size: 11px !important;
+    border-radius: 16px !important;
+  }
+
+  .sender-name {
+    max-width: 110px !important;
+  }
+
+  .btn-change-settings {
+    padding: 6px 10px !important;
+    font-size: 11px !important;
+  }
+
+  .quick-date-chips-wrapper {
+    display: grid !important;
+    grid-template-columns: repeat(3, 1fr) !important;
+    gap: 6px !important;
+  }
+
+  .quick-date-chip {
+    height: 38px !important;
+    font-size: 11px !important;
+    padding: 0 4px !important;
+    min-width: 0 !important;
+    width: 100% !important;
+    border-radius: 9px !important;
+  }
+
+  .btn-save-settings {
+    height: 42px !important;
+    font-size: 12px !important;
+    border-radius: 10px !important;
+  }
+
+  .btn-add-draft-modern {
+    height: 44px !important;
+    font-size: 12px !important;
+    border-radius: 10px !important;
+  }
+
+  .ukm-chip {
+    padding: 5px 10px !important;
+    font-size: 11px !important;
+  }
+
+  .photo-thumbnail-grid {
+    grid-template-columns: repeat(auto-fill, minmax(100px, 1fr)) !important;
+    gap: 8px !important;
+  }
+
+  .photo-thumbnail-img-wrapper {
+    height: 75px !important;
+  }
+
+  .dropzone-inner {
+    gap: 10px !important;
+  }
+
+  .dropzone-icon-circle {
+    width: 36px !important;
+    height: 36px !important;
+  }
+
+  .dropzone-main-text {
+    font-size: 11px !important;
+  }
+
+  .hide-on-mobile { display: none !important; }
+}
