+++ src/features/report/styles/form-shared.css
@@ -1,0 +1,144 @@
+/* Section Cards & Containers */
+.global-section,
+.active-form-section,
+.draft-list-wrapper {
+  background: var(--color-bg-section);
+  border: 1px solid var(--color-border-subtle);
+  border-radius: var(--radius-lg);
+  padding: 22px;
+  box-shadow: 0 4px 20px rgba(88, 28, 135, 0.04);
+  margin-bottom: 24px;
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
+
+/* Inputs & Labels */
+.input-group {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+}
+
+.form-label {
+  color: #334155;
+  text-align: left;
+  font-size: 12px;
+  font-weight: 800;
+  letter-spacing: 0.5px;
+}
+
+.form-input {
+  width: 100%;
+  background: var(--color-bg-input);
+  border: 1.5px solid var(--color-border-default);
+  color: var(--color-text-main);
+  border-radius: var(--radius-md);
+  height: 48px;
+  font-size: 14px;
+  font-weight: 600;
+  padding: 0 16px;
+  font-family: var(--font-family-base);
+  transition: all 0.2s ease;
+  box-sizing: border-box;
+}
+
+.form-input:focus {
+  outline: none;
+  border-color: var(--color-primary);
+  box-shadow: 0 0 0 4px rgba(217, 70, 239, 0.15);
+  background: #ffffff;
+}
+
+/* Custom Dropdown */
+.custom-select-wrapper {
+  position: relative;
+  width: 100%;
+}
+
+.custom-select-trigger {
+  width: 100%;
+  background: var(--color-bg-input);
+  border: 1.5px solid var(--color-border-default);
+  color: var(--color-text-main);
+  border-radius: var(--radius-md);
+  height: 48px;
+  font-size: 14px;
+  font-weight: 600;
+  padding: 0 16px;
+  font-family: var(--font-family-base);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  user-select: none;
+  box-sizing: border-box;
+}
+
+.custom-select-trigger.active {
+  border-color: var(--color-primary);
+  box-shadow: 0 0 0 4px rgba(217, 70, 239, 0.15);
+  background: #ffffff;
+}
+
+.custom-select-trigger.disabled {
+  background-color: #f1f5f9;
+  color: #94a3b8;
+  border-color: #e2e8f0;
+  cursor: not-allowed;
+}
+
+.custom-select-value {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+}
+
+.custom-select-placeholder {
+  color: var(--color-text-muted);
+}
+
+.chevron-icon {
+  font-size: 11px;
+  color: var(--color-text-muted);
+  transition: transform 0.3s ease;
+}
+
+.custom-select-trigger.active .chevron-icon {
+  transform: rotate(180deg);
+}
+
+.custom-select-dropdown {
+  position: absolute;
+  top: calc(100% + 6px);
+  left: 0;
+  right: 0;
+  background: #ffffff;
+  border: 1px solid var(--color-border-default);
+  border-radius: 14px;
+  box-shadow: var(--shadow-dropdown);
+  z-index: 50;
+  max-height: 240px;
+  overflow-y: auto;
+  animation: slideDown 0.2s cubic-bezier(0.16, 1, 0.3, 1);
+  padding: 6px 0;
+}
+
+.custom-select-item {
+  padding: 10px 16px;
+  font-size: 13px;
+  font-weight: 600;
+  color: #475569;
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  cursor: pointer;
+  transition: all 0.15s ease;
+}
+
+.custom-select-item:hover,
+.custom-select-item.selected {
+  background: rgba(217, 70, 239, 0.08);
+  color: #c026d3;
+}
